Update projectroadmap.md
I added my sprint 2 updated comments regarding the UI code repo and a Pomodoro for time blindness, they are too complex for the time remaining for this project and they have many files that would all need to integrate and import many modules, I have pivoted to enhance the existing UI and task list we have as a scope change. I also added the tasks I have in the roadmap to add functionality to mark tasks complete, count total tasks completed, and add the button and voice commands that go along with completion and positive reinforcement.