Update OrbitReminder.py
I added a way show the task is selected in the list and to visually indicate task completion. I also added audio messages to the have the voice in the app to say the task is completed and tells the user the total tasks completed thus far that day for positive reinforcement.